Label views can now be selected.
When a node in the hyper view is selected, the labelView for that node
is placed at the end of the canvasItems list ( so as to be drawn last ).
Also the LabelView transparency is turned off while the current state of
the hyper view is unchanged.
Also made some miner clean-ups.