Description panel is set to clear all details on each new query.
Description panel is setting all details to the root node on each new query and on the startup.

Bug fixed in clones panel: when a user first time clicks on a clone in one of the views - clones panel doesn't get list of clones (hyperview and tree view HAVE to be initialised BEFORE description panel).

Rdf parser tries to strip carriage returns and whitespaces from rdf objects.