WebKB2Source class looks to see if a query is ambugious.
A query that is ambugious has more then one RDF document returned
in the input stream. This class will find the root term in each document and
return the list of terms to the user to choose which term they ment.
i.e "cat" can be (big_cat, Caterpillar, true_cat, etc)