Fixed a problem with unconnected nodes list - was getting synonyms and descriptions in there.
Outstanding issues:
introduced a bug - if a top level node is in the signature of relaton types,
we don't add it to the list of unconnected nodes, although we should.
This bug is apparent in the CoreCommunication ontology - keep getting different
number of unconnected nodes. Probably need a notion of RelationNode rather then
using node types to distinguish between relation and concept types.