Removed checking if source is dynamic. Added parameter to xml file (see descr. below)
Now app behaves in the same way for static and dynamic ontologies

Added parameter for edges displayed in description view: isQueryOn. If this is
true - then description panel will enable a user to query for this item using
hyperlink-like label. This enables us to not hardcode supertype relationship
in the code.