Added capability to import SVG shapes.
An import function allows importing simple SVG shapes (all styles ignored),
the OntoramaConfig class tries to find a file "node.svg" in the default config
location and imports this as concept shape if found, otherwise we still use
the airplane.

One way to create these files easily is using StarOffice/OpenOffice: create
a drawing, insert a clipart object and export as SVG. Other drawing objects
haven't been tested.

Problem: for some reason the SAXBuilder insists on loading the DTD if the
DOCTYPE is given. This is extra nasty if there is no connecion to the internet
or a firewall blocks access to the DTD, since it will cause an Exception after
a timeout period. Workaround: removing the DOCTYPE declaration.