Code cleanup: preparation for layers in canvas.
Some code accessing the canvasItems member from the Canvas is
replaced with nicer code, using access methods from the canvas. The new
code will work with the layers that are about to be introduced in the
Tockit Canvas library, the old one wouldn't. Readability should be a bit
better, too.