Trying to change the way query works.
So far - changed QueryEngine to accept Source, Parser, etc as parameters (as strings for now).
To do that had to pass this information withing the Query (should change that).
This should allow to cross-load data between different backends.

Added method createGraph() to Backend interface - so correct graph implementation is created.
This has major drawback - query enging has to ask configuration for current backend.