changed TestCase for P2P writer
Will not  write a model and parse it using P2P parser in order to figure
out if we wrote it correctly. For the moment will write a model and read
RDF model back and compare with the expected model. However, there 
should be a better solution since this will not give us descriptive enough
message in case of error.