Renamed jxta.jar into jxtaOntoRama.jar.
Reason: found a cause of the bug when peers can't find each other - it
is apparent that the reason for this is when peers are run from different
environments (eclipse and cygwin/ant or eclipse and dos) they use different
jxta jars (not sure what default order is for JVM). That in itself shouldn't be 
a problem, but we noticed that using different jars leads to different 
program behaviour in terms of default global group.

So here is the problem: jxta.jar has been recompiled and edited so it
is not the default jxta.jar obtained from http://jxta.org. To address this
we will rename jxta.jar into jxtaOntoRama.jar so there is no confusion.

Outstanding issues:
- find out if it is possible to reset default net group in some other way 
rather then hacking jxta lib.