Fixed a bug
When building a query from query panel or requery some part of a graph - 
was getting a qraph without edge types displayed in the description panel - 
they were displayed as unconnected nodes. This was due to us building
a query with only edge types displayed in the query panel and not including
edge types displayed in the description panel. This is fixed now.