Fixed a bug: history menu didn't work.
Cause: when switched to using backends, we no longer keep track of examples in
history menu which means we don't have access to each example's confitg 
settings - instead backend just uses current settings when executing a 
query. 

Solution: not quite happy with this solution, but it seems it works at least for ExamplesBackend.

- added a method to backend interface: getQueryEngine that returns 
a QueryEngine used for last query.

- HistoryElement will now keep track of a QueryEngine corresponding to 
it's Query.

- introduced overloaded constructor to QueryEngineThread that allows
us to exectute queries for a given query engine as well as passing queries
to a current backend.

- added new event: HistoryQueryStartEvent and corresponding handler.