Attempting to fix a problem when running WebKB interface example.
When searching for a 'dog' - get a looooong error message that 
is too wide for the screen.

Solution:
trying to add some carriage returns in the error message string.
This has been partly successfull: we break the string up by spaces and 
trying to compose a number of smaller length strings - the result is 
not always easy to read because our carriage returns are not placed semantically.