Fixed (hacked) problem with history.
Problem: 
 when we displayed a history item and then we want to search 
for a term in the displayed ontology we get error: term not found in
the result set.

Cause:  query engine config should be reset to the current config in the backend.
and in ExamplesBackend case - we should reset currently active example

Solution: added setQueryEngine() method to Backend interface and 
implementing classes.