Some refactorings:  now we have a method to search for peers in global Net Group.
Reason for this is: we can't easily check if we can see other peers unless 
we join the same group they are in. The problem with this is that we often
can't see other groups at all, so we can't join them, so we can't see other
peers.

Changed method Communication.getGlobalPG to static so we dont' need to
drag reference for the object everywhere. Probably need some sort of 
class that can remember all globals.

naturally, there is a lot more debugging output now ;)