Attempts to fix resizing problem in peers panel.
Problem: list containing peers resizes on repaint. if peer names are
short - we get skinny ugly list.

Solution: had to use setPreferredSize method to controll size of the 
scroll panel containing list.

Issues:

- had to hardcode size we are setting. All attempts to figure out how
to calculate it depending on other components failed: we only can get
size of these components after we added all children components. Since
our component we are trying to resize is a child - we  initialise it before 
swing calculates size of a parent. Can't see a way to overcome it except
a string of hacks.

- another possiblity to controll size of the list is to use setCellPrototype 
method on a jlist. This proved unsatisfactory solution because when
its set - it kills scroll bar and trancates long cell entries to the length of
the prototype cell.