Table Cell Renderer is working correctly.
To solve the problems all it took is to realise that each renderer works on
column Class, you can't set it for certain columns as I assumed and you 
also can't set one renderer for the whole table if it has different classes
in different columns. What one needs to do is to set one renderer per 
each class used in the table (if default renderer can't be used for these).