Fixed some issues with URI's for asserters and rejectors.
However, it is not clear to me how this is going to work:

- assertions: if we added a node, we add an assertion for it and keep
it in our model (how do we get assertion information from other peers?)

- rejections: if we deleted a node, we remove it add a rejection for it. 
Since it is not in our model anymore - how do we ever access this info?
or how will other peers keep this info? once it's propagated - it's gone?