Changed test expectations to match results.
This time I didn't dig in too deeply, but the results seem consistently off and since the other test failed with similar offsets it seems pretty plausible that the test is broken due to some change in the input file, which I don't want to investigate right now. It should be good enough to cover us from breaking things while working on the production code.