Another round of introducing generics.
 * let Eclipse infer generic arguments
 * fix things it broke
 * update other bits along the fly
 
Bigger item this time: the identity of EdgeType isn't well defined. It is used as key in hash structures, but the _namespace member gets changed after that happens, so using the namespace as part of equals(..)/hashCode() causes bugs. It probably should be used, though. Quite possibly it should be one namespace per direction, too.