Added basic support for OWL files.
The file backend can now handle OWL files using the OWLAPI library. Currently only subclassing is supported as relation and querying doesn't work either. The hierarchy displayed will be one of the tallest hierarchies in the file, there is no way to change over if there are multiple independent hierarchies.